Aula 07-11-2022
- CRUD produtos (upload de imagens)
- Controle de acesso aos controllers (configurado app\config\Filters.php e criado a classe Filters.php na pasta app\Filters)
- Página home com produtos sendo apresentados conforme cadastrados na base de dados.

+++ app/Views/admin/listaProduto.php
@@ -19,19 +19,31 @@
                     </tr>
                 </thead>
                 <tbody>
-                    <?php foreach ($data as $value): ?>
+
+                    <?php if (count($data) > 0): ?>
+
+                        <?php foreach ($data as $value): ?>
+                            <tr>                    
+                                <td><?= $value['descricao'] ?></td>
+                                <td><?= $value['departamentoDescricao'] ?></td>
+                                <td><?= formatValor($value['precoVenda']) ?></td>
+                                <td><?= mostraStatus($value['statusRegistro']) ?></td>
+                                <td>
+                                    <a href="<?= base_url() ?>/Produto/form/view/<?= $value['id'] ?>" class="btn btn-secondary btn-sm btn-icons-crud" title="Visualizar"><i class="fa fa-eye" aria-hidden="true"></i></a>    
+                                    <a href="<?= base_url() ?>/Produto/form/update/<?= $value['id'] ?>" class="btn btn-secondary btn-sm btn-icons-crud" title="Alterar"><i class="fa fa-file" aria-hidden="true"></i></a>    
+                                    <a href="<?= base_url() ?>/Produto/form/delete/<?= $value['id'] ?>" class="btn btn-secondary btn-sm btn-icons-crud" title="Excluir"><i class="fa fa-trash" aria-hidden="true"></i></a>                               
+                                </td>
+                            </tr>
+                        <?php endforeach; ?>
+
+                    <?php else: ?>
+
                         <tr>                    
-                            <td><?= $value['descricao'] ?></td>
-                            <td><?= $value['departamentoDescricao'] ?></td>
-                            <td><?= formatValor($value['precoVenda']) ?></td>
-                            <td><?= mostraStatus($value['statusRegistro']) ?></td>
-                            <td>
-                                <a href="<?= base_url() ?>/Produto/form/view/<?= $value['id'] ?>" class="btn btn-secondary btn-sm btn-icons-crud" title="Visualizar"><i class="fa fa-eye" aria-hidden="true"></i></a>    
-                                <a href="<?= base_url() ?>/Produto/form/update/<?= $value['id'] ?>" class="btn btn-secondary btn-sm btn-icons-crud" title="Alterar"><i class="fa fa-file" aria-hidden="true"></i></a>    
-                                <a href="<?= base_url() ?>/Produto/form/delete/<?= $value['id'] ?>" class="btn btn-secondary btn-sm btn-icons-crud" title="Excluir"><i class="fa fa-trash" aria-hidden="true"></i></a>                               
-                            </td>
+                            <td colspan="5">Nenhum produto cadastro no momento...</td>
                         </tr>
-                    <?php endforeach; ?>                
+
+                    <?php endif; ?>
+
                 </tbody>
             </table>
         </div>
